pyproject.toml
[tool.poetry]
name = "grimwaves-api"
version = "0.3.1"
description = "FastAPI service that provides endpoints for servicing the GrimWaves channel"
authors = ["Denis Zwinger"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
tomli = "^2.2.1"
pydantic-settings = "^2.8.1"
celery = "^5.5.0"
redis = "^5.2.1"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
aiohttp = "^3.11.16"
h11 = "^0.16.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
mypy = "^1.15.0"
celery-types = "^0.23.0"
hvac = "^2.3.0"
types-hvac = "^2.3.0.20240621"
safety = "^3.4.0"
bandit = "^1.8.3"
basedpyright = "^1.29.1"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
types-psutil = "^7.0.0.20250401"
pytest-mock = "^3.14.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
lint.select = ["E", "F"]
lint.ignore = [
    "E501",     # line-too-long 	                            Line too long (120 > 110 characters)
    "D1",       # undocumented-public-module 	                Missing docstring in public module
    "D203",     # one-blank-line-before-class 	                1 blank line required before class docstring
    "D213",     # multi-line-summary-second-line 	            Multi-line docstring summary should start at the second line
    "FA102",    # future-required-type-annotation 	            Missing from __future__ import annotations, but uses
    "ANN401",   # any-type 	                                    Dynamically typed expressions (typing.Any) are disallowed in {name}
    "FBT001",   # boolean-type-hint-positional-argument 	    Boolean-typed positional argument in function definition
    "FBT002",   # boolean-default-value-positional-argument 	Boolean default positional argument in function definition
    "TRY301",   # raise-within-try 	                            Abstract raise to an inner function
    "T201",     # print 	                                    print found
    "PLR0913"   # too-many-arguments 	                        Too many arguments in function definition ({c_args} > {max_args})
]
cache-dir = "/tmp/ruff-cache"
lint.pydocstyle.convention = "google"
exclude = [".venv", "docs", "backups"]

[tool.ruff.lint.isort]
no-lines-before = ["standard-library", "local-folder"]
known-third-party = []
known-local-folder = ["whole_app"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*.py" = ["ANN401", "S101", "S311", "ANN201", "ANN001"]

[tool.basedpyright]
include = ["."]
exclude = ["backups", ".venv", "docs"]
reportAny = false
reportPrivateLocalImportUsage = false
reportExplicitAny = false

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
norecursedirs = ["backups", ".venv", "docs"]
markers = [
    "memory: marks tests as memory intensive to run.",
    "integration: marks integration tests.",
    "stress: marks stress tests.",
    "e2e: marks tests as end-to-end tests",
]

[tool.safety]
ignore_paths = ["backups/"]

[tool.mypy]
exclude = "backups/"
